## Summary

Improved the speed of `Shuffle()` by ～ 2.6x.

## Benchmark

```cs
public static int Main(string[] args)
{
    BenchmarkRunner.Run<ShuffleBench>(args: args);
    return 0;
}
```

```
BenchmarkDotNet v0.15.2, Windows 11 (10.0.26100.4349/24H2/2024Update/HudsonValley)
12th Gen Intel Core i7-12700F 2.10GHz, 1 CPU, 20 logical and 12 physical cores
.NET SDK 10.0.100-preview.5.25277.114
  [Host]     : .NET 10.0.0 (10.0.25.27814), X64 RyuJIT AVX2
  DefaultJob : .NET 10.0.0 (10.0.25.27814), X64 RyuJIT AVX2
```

Before:

| Method          | N      | M      | Mean          | Error         | StdDev         | Median        |
|---------------- |------- |------- |--------------:|--------------:|---------------:|--------------:|
| ShuffleTakeLinq | 1000   | 10     |   4,390.91 ns |    256.749 ns |     719.953 ns |   4,670.29 ns |
| ShuffleTakeLinq | 1000   | 100    |   7,343.62 ns |    345.180 ns |     984.817 ns |   7,719.69 ns |
| ShuffleTakeLinq | 1000   | 1000   |   4,071.74 ns |     53.117 ns |      41.470 ns |   4,061.72 ns |
| ShuffleTakeLinq | 1000   | 100000 |   4,117.56 ns |     33.945 ns |      26.502 ns |   4,124.88 ns |
| ShuffleTakeLinq | 100000 | 10     | 327,208.70 ns | 16,008.675 ns |  47,201.941 ns | 343,945.87 ns |
| ShuffleTakeLinq | 100000 | 100    | 242,080.14 ns | 26,929.140 ns |  79,401.182 ns | 186,970.47 ns |
| ShuffleTakeLinq | 100000 | 1000   | 229,417.80 ns |  3,317.590 ns |   8,200.262 ns | 227,635.45 ns |
| ShuffleTakeLinq | 100000 | 100000 | 630,294.22 ns | 65,411.969 ns | 192,868.680 ns | 512,984.18 ns |
|                 |        |        |               |               |                |               |
| ShuffleTakeZL   | 1000   | 10     |      82.69 ns |      0.704 ns |       0.658 ns |      82.69 ns |
| ShuffleTakeZL   | 1000   | 100    |     690.60 ns |     36.599 ns |     107.914 ns |     724.73 ns |
| ShuffleTakeZL   | 1000   | 1000   |   6,037.47 ns |    369.983 ns |   1,090.904 ns |   6,394.99 ns |
| ShuffleTakeZL   | 1000   | 100000 |   5,672.11 ns |    329.154 ns |     965.351 ns |   5,971.55 ns |
| ShuffleTakeZL   | 100000 | 10     |  11,081.71 ns |    485.221 ns |   1,344.546 ns |  11,431.81 ns |
| ShuffleTakeZL   | 100000 | 100    |  11,825.45 ns |    508.941 ns |   1,418.723 ns |  12,206.31 ns |
| ShuffleTakeZL   | 100000 | 1000   |  17,106.84 ns |    962.291 ns |   2,760.994 ns |  18,068.87 ns |
| ShuffleTakeZL   | 100000 | 100000 | 258,360.04 ns |  2,983.644 ns |   2,329.433 ns | 258,010.50 ns |

After:

| Method          | N      | M      | Mean          | Error        | StdDev       | Median        |
|---------------- |------- |------- |--------------:|-------------:|-------------:|--------------:|
| ShuffleTakeLinq | 1000   | 10     |   4,398.77 ns |   237.561 ns |   692.976 ns |   4,664.29 ns |
| ShuffleTakeLinq | 1000   | 100    |   6,158.41 ns |   531.077 ns | 1,565.892 ns |   5,631.29 ns |
| ShuffleTakeLinq | 1000   | 1000   |   4,043.41 ns |    20.010 ns |    18.717 ns |   4,038.13 ns |
| ShuffleTakeLinq | 1000   | 100000 |   4,092.15 ns |    33.186 ns |    31.042 ns |   4,083.33 ns |
| ShuffleTakeLinq | 100000 | 10     | 176,712.14 ns | 1,863.189 ns | 1,742.828 ns | 176,204.79 ns |
| ShuffleTakeLinq | 100000 | 100    | 185,034.66 ns |   846.611 ns |   750.498 ns | 184,857.79 ns |
| ShuffleTakeLinq | 100000 | 1000   | 227,024.46 ns | 1,310.362 ns | 1,225.713 ns | 226,563.96 ns |
| ShuffleTakeLinq | 100000 | 100000 | 491,479.11 ns | 4,557.176 ns | 4,262.785 ns | 490,562.30 ns |
|                 |        |        |               |              |              |               |
| ShuffleTakeZL   | 1000   | 10     |      81.37 ns |     1.307 ns |     1.222 ns |      81.69 ns |
| ShuffleTakeZL   | 1000   | 100    |     265.07 ns |     2.620 ns |     2.451 ns |     264.80 ns |
| ShuffleTakeZL   | 1000   | 1000   |   1,921.05 ns |    15.893 ns |    14.089 ns |   1,919.34 ns |
| ShuffleTakeZL   | 1000   | 100000 |   1,982.34 ns |    18.037 ns |    16.871 ns |   1,980.04 ns |
| ShuffleTakeZL   | 100000 | 10     |   6,535.99 ns |    49.108 ns |    45.936 ns |   6,525.33 ns |
| ShuffleTakeZL   | 100000 | 100    |   6,717.00 ns |    40.599 ns |    35.990 ns |   6,708.83 ns |
| ShuffleTakeZL   | 100000 | 1000   |   8,749.90 ns |    77.955 ns |    72.919 ns |   8,749.02 ns |
| ShuffleTakeZL   | 100000 | 100000 | 225,266.30 ns | 1,117.694 ns |   990.806 ns | 225,655.81 ns |

## Details 

Algorithm details:

* https://github.com/dotnet/runtime/pull/111015
* (Japanese) https://andantesoft.hatenablog.com/entry/2024/12/30/183006

I implemented the PRNG without relying on `System.Random` because the algorithm requires `NextUInt64()`.
The algorithm of the PRNG itself has not changed. It uses `xoshiro256**`, the same as the current .NET implementation.
In environments with older implementations of `System.Random`, such as Unity, this may help speed things up even more.